fix: ignore unprintables for truncation with suffix
Previously if a progress line was so long that it would be longer than
100 characters (which is currently used as a terminal width approximation I suppose),
truncating the buffer and replacing the last 4 characters with the ellipsis
would look inconsistent. This was because overflows were detected by
catching `error.NoSpaceLeft` on buffer writing which didn't work well for
unprintable characters such as escape sequences.

This is fixed by using a more sophisticated approach using buffered and counting writers.
At the same time this also serves as a modernization of the code.

The issue described only happened on non-Windows systems.