docs: update doc comments
This updates the documentation comments to no longer say that these functions or structs
are intended for use only within test blocks because that is now automatically ensured, so there is no need to tell the user about it.
If the user does, they'll get an error.

Apart from that I also fixed some other minor things I noticed while reading through the docs.