feat: improve checkScope error
This is supposed to improve the error of checkScope when a thing was not inside a test block.
This however doesn't work right now and fails with no error return trace so it's hard to investigate.

Let's maybe reuse this code in a month or so when the compiler actually tells us what's wrong with this code.