link: introducing the PRG output format
This adds a new output format and linker usable with -ofmt=prg.
The PRG output format can more or less be called the native executable format
of a lot of Commodore machines, including the Commodore 64.
This linker is still in a very experimental stage.
It has currently only been tested with the C64 but it is supposed to be
compatible with more of the Commodore family later on.