prg: fix memory leaks
We were incorrectly simply taking the ArrayList as a slice, ignoring
any additional capacity the ArrayList might have had allocated.
This was causing problems particularly with big code blocks where
the length of the block would diverge from the capacity of the
ArrayList.
Now with toOwnedSlice() we deallocate any excessive memory.