6502: forget about branch instruction optimizations for now
I was making efforts to allow emitting a branch instruction (such as
BVC) instead of a jump instruction (JMP) to save a single byte of code
size when the body code size is <= 128 but I'm burying that effort for now because the gain of this
optimization is very minor.
In order for it to work we have to choose a branch instruction that we
know will always branch but in the beginning of a function's codegen
we don't know if any flags are set so we have to set one ourselves at
least once and also determine whether it's worth doing the optimization in the first place
for which we would have to know how many times we have a certain construct in our
function's AIR (or even better the AIR of the whole codegen)
which also makes this hard.

All in all, this seemed like an easy optimization but is actually quite
involved.